fix send token flow
